+++ NEWS.md
@@ -1,3 +1,4 @@
 # parsnip 0.0.0.9000
 
+* The `fit` interface was previously used to cover both the x/y interface as well as the formula interface. Now, `fit` is the formula interface and [`fit_xy` is for the x/y interface](https://github.com/topepo/parsnip/issues/33). 
 * Added a `NEWS.md` file to track changes to the package.